requirements.txt
requests
python-dotenv
selenium==4.25.0
typing_extensions>=4.12.2
